Refactor rating and test code for better readability
The Rating and associated Test classes have been relocated from `Http` namespace to the root level for a better organisation, as they are not strictly HTTP related. Usage of the DB Facade in test cases has been replaced by using Model instance for a more Eloquent approach, improving readability and making use of Laravel's ORM. User authentication check is added after acting as a user in tests for a more comprehensive test case.